+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Use stdout as output if "-" is given as file ([#10](https://github.com/orhun/menyoki/issues/10))
+
 ### Changed
 - Allow disabling action keys on slop selection ([#12](https://github.com/orhun/menyoki/issues/12))